Added separate Modal component
- cleared up styles
